// agile-benefit-tracker/server/src/models/enums.ts
export enum ClaimType {
  OPD = 'OPD',
  WELLNESS = 'Wellness',
}

export enum ClaimStatus {
  SUBMITTED = 'Submitted',
  APPROVED = 'Approved',
  REJECTED = 'Rejected',
  AUTO_REJECTED = 'Auto-Rejected',
  PAID = 'Paid',
}

export enum UserRole {
  EMPLOYEE = 'Employee',
  HR = 'HR',
  ACCOUNTS = 'Accounts',
}